Split ProcessTest and add more tests to shared
