+++ unit-tests/native/src/test/scala-2.11/scala/Issue2305.scala
