+++ unit-tests/shared/src/test/scala/EmptyPackageTest.scala
