+++ unit-tests/shared/src/test/scala/javalib/lang/ProcessTest.scala
@@ -1,12 +1,9 @@
-package java.lang
+package javalib.lang
 
 import java.util.concurrent.TimeUnit
 import java.io._
 import java.nio.file.Files
 
-import scala.scalanative.unsigned._
-import scala.scalanative.unsafe._
-import scala.scalanative.posix.{fcntl, unistd}
 import scala.io.Source
 
 import org.junit.Test
@@ -15,60 +12,11 @@ import org.junit.Assert._
 class ProcessTest {
   import javalib.lang.ProcessUtils._
 
-  private def assertProcessExitOrTimeout(process: Process): Unit = {
-    // Suspend execution of the test until either the specified
-    // process has exited or a reasonable wait period has timed out.
-    //
-    // A waitFor() prevents zombie processes and makes the exit value
-    // available. A timed waitfor means the test will eventually complete,
-    // even if there is a problem with the underlying process.
-    //
-    // In the normal case, the process will exit within milliseconds or less.
-    // The timeout will not increase the expected execution time of the test.
-    //
-    // Ten seconds is an order of magnitude guess for a "reasonable"
-    // completion time.  If a process expected to exit in milliseconds
-    // takes that three orders of magnitude longer, it must be reported.
-
-    val tmo = 10
-    val tmUnit = TimeUnit.SECONDS
-
-    assertTrue(
-      s"Process took more than $tmo ${tmUnit.name} to exit.",
-      process.waitFor(tmo, tmUnit)
-    )
-  }
-
-  val scripts = Set("echo.sh", "err.sh", "ls", "hello.sh")
-
   @Test def ls(): Unit = {
     val proc = new ProcessBuilder("ls", resourceDir).start()
     val out = readInputStream(proc.getInputStream)
 
     assertProcessExitOrTimeout(proc)
-
-    assertEquals(scripts, out.split("\n").toSet)
-  }
-
-  @Test def inherit(): Unit = {
-    val f = Files.createTempFile("/tmp", "out")
-    val savedFD = unistd.dup(unistd.STDOUT_FILENO)
-    val flags = fcntl.O_RDWR | fcntl.O_TRUNC | fcntl.O_CREAT
-    val fd = Zone { implicit z =>
-      fcntl.open(toCString(f.toAbsolutePath.toString), flags, 0.toUInt)
-    }
-
-    val out =
-      try {
-        unistd.dup2(fd, unistd.STDOUT_FILENO)
-        unistd.close(fd)
-        val proc = new ProcessBuilder("ls", resourceDir).inheritIO().start()
-        proc.waitFor(5, TimeUnit.SECONDS)
-        readInputStream(new FileInputStream(f.toFile))
-      } finally {
-        unistd.dup2(savedFD, unistd.STDOUT_FILENO)
-        unistd.close(savedFD)
-      }
 
     assertEquals(scripts, out.split("\n").toSet)
   }
