+++ unit-tests/shared/src/test/scala/scala/ArrayDoubleCopyTest.scala
