+++ unit-tests/shared/src/test/scala/scala/ArrayObjectCopyTest.scala
