+++ unit-tests/shared/src/test/scala/scala/DivisionByZeroTest.scala
