+++ unit-tests/shared/src/test/scala/scala/DoubleComparisonTest.scala
