+++ unit-tests/shared/src/test/scala/scala/EnumerationTest.scala
