+++ unit-tests/shared/src/test/scala/scala/ExceptionHandlingTest.scala
