+++ unit-tests/shared/src/test/scala/scala/FloatComparisonTest.scala
