+++ unit-tests/shared/src/test/scala/scala/FloatingToIntegerOverflowTest.scala
