+++ unit-tests/shared/src/test/scala/scala/IsInstanceOfTest.scala
