+++ unit-tests/shared/src/test/scala/scala/NullPointerTest.scala
