+++ unit-tests/shared/src/test/scala/scala/ObjectCloneTest.scala
