+++ unit-tests/shared/src/test/scala/scala/ReflectiveProxyTest.scala
