+++ unit-tests/shared/src/test/scala/scala/ShiftOverflowTest.scala
