+++ unit-tests/shared/src/test/scala/scala/SymbolTest.scala
