+++ unit-tests/shared/src/test/scala/scala/bugcompat/LongFloatPrimitiveTest.scala
