+++ unit-tests/shared/src/test/scala/scala/compiler/DefaultMethodsTest.scala
