+++ unit-tests/shared/src/test/scala/scala/partest/DigitIdentifiersTest.scala
