+++ unit-tests/shared/src/test/scala/scala/partest/MixinTest.scala
