+++ unit-tests/shared/src/test/scala/scala/util/PropertiesTest.scala
